requirements.txt
#
# ===================================================================================
#  The Shopping List: requirements.txt
# ===================================================================================
#
discord.py
Flask
requests
python-dotenv
gunicorn